[REFACTOR] Create a Blake2bHash type
It's nice to make things explicit when we can!

Rather than having client code have to keep track of different
operations on these bytes, we can just make a wrapper that
consolidates all that behavior into a single place. It'll make call
sites more explicit and cleaner.

One change in particular is that we're now doing constant-time
comparison of hashes. This is most likely overkill, strictly speaking,
since the comparison isn't doing anything with secret material.

Signed-off-by: Christopher Maier <cmaier@chef.io>