[REFACTOR] All the "hash" functions return Blake2bHash, not String
This continues to unify around the `Blake2bHash` type.

One thing *not* tackled in this is a unification of the various "write
a file if its hash has changed from what's on disk" calls. We have 3
instances of this logic, all of which are subtly different in a number
of ways. That should all be unified, but for now, we only change as
much as is needed to move from `String` to `Blake2bHash`.

Signed-off-by: Christopher Maier <cmaier@chef.io>