[REFACTOR] Add various keypair storage functions to the KeyCache
This provides a central place where "save a keypair" logic can
live. Long-term, it would be nice to make this an all-or-nothing
operation, where either both keys are written, or neither are. I think
that will require a little more work to clean up our error handling in
this crate. In any event, the previous implementation did *not* behave
in an all-or-nothing way, so this isn't changing any behavior.

Signed-off-by: Christopher Maier <cmaier@chef.io>