additional safety bounds for ident validity check
This will prevent bad PackageIdent version or release fields from
creating an inoperable situation on either the filesystem or via HTTP
calls to Builder.

In both cases, sequences such as '/../' or '/./' or even '/..1/' in the package name
create problems for operations on the CLI or Builder API as those have
special meaning and result in an inoperable or unintended outcome.

Signed-off-by: Jeremy J. Miller <jm@chef.io>