[REFACTOR] Pull key generation interface into KeyCache
In real code, anytime you create a key, you'll always need to save it
to disk. As such, we can make life a little easier on people by
pulling that repetetive code into the KeyCache struct, making it the
central hub for all key operations of creation, storage, and
retrieval.

It also has the nice side effect of simplifying some test code, so
that's cool, too.

Signed-off-by: Christopher Maier <cmaier@chef.io>