[REFACTOR] Ensure Builder can use Blake2bHash type
Builder needs to be able to dereference a hash as a `&str`, so we add
the hex string to the Blake2bHash so we can implement
`Deref<Target=str>`.

We also needed to change the artifact checksum type back to String for
now.

Once Builder has been updated to use these new types, we can come back
and simplify things.

Signed-off-by: Christopher Maier <cmaier@chef.io>