Add a Builder encryption key
This represents the encryption logic used by Builder to encrypt secret
information at rest. While it shares the same underlying machinery as
the rest of our encryption keys, it has a distinct usage pattern in
Builder.

In particular, all encryption is a signed box where sender and
recipient are the same. This means that we know that all encrypted
content was encrypted by Builder, and only Builder can decrypt
it. This is reflected in the API for the `BuilderSecretEncryptionKey`.

As a consequence of this, a corresponding public key is not actually
needed. As a result, there is no type to represent it. Thus, this is
the only encryption key in Habitat that does not exist as a pair.

Signed-off-by: Christopher Maier <cmaier@chef.io>