[REFACTOR] Replace libsodium_sys with blake2b_simd for hashing
We not only get to replace a lot of complex, custom code, but get to
eliminate a lot of unsafe code as well.

Since everything else in the codebase interacts with our `Blake2bHash`
wrapper type, nothing else actually has to change!

(However, the `hash_string` function is no longer necessary, since it
can be done by `hash_bytes`.)

Signed-off-by: Christopher Maier <cmaier@chef.io>