[REFACTOR] Utilize the `Origin` type more broadly
There's no guarantee that this covers *all* uses of origins in the
codebase. Rather, I started with the key-related interfaces, switching
all `String` or `&str` uses that were really Origin names into
`Origin` types. This ended up cascading out to a number of other
places, which are all addressed in this PR.

Mainly, this amounted to a lot of type switches, with a handful of
less-mechanical changes.

* Add an `AsRef<str>` implementation to `Origin`
* Add `hacky_get_origin` to `PackageIdent`
  Ultimately, we'll store an `Origin` in the `PackageIdent` type, at
  which point, we'll return `&Origin` from the `origin()`
  method. Until then, this intentionally-named method is used to
  centralize the "parse-and-expect" logic of converting the internal
  `String` into an `Origin`. We will, of course, remove this method
  soon :)
* The logic for populating a `hab pkg build` invocation with secret
  signing keys has been updated. In particular, the logic for creating
  the comma-delimited string we pass to the executable is pushed closer
  to the invocation site.

Signed-off-by: Christopher Maier <cmaier@chef.io>