[CI] Capture Supervisor logs from e2e tests
When e2e tests fail, it's often really nice to have access to the logs
of the Supervisor to help debug things. Unfortunately, we had no way
to do that previously.

Some tests captured Supervisor output to a named file (like
"sup.log"), while others captured it to a randomly-named temporary
file. The former is bad because some test suites start up the
Supervisor multiple times (thus clobbering logs of any prior
Supervisor processes), and the latter is bad because we never knew
what the name of the file actually was. And both cases were bad
because Buildkite didn't know how to capture either of them!

Here, we add a custom function to generate a log file name. It takes a
label, used to distinguish different Supervisor runs within a test
suite, and appends a random suffix, along with the `.log`
extension. All uses of the `Start-Supervisor` function (as well as
some other functions) are modified to use this filename function for
logs. Finally, Buildkite has been instructed to capture all `*.log`
files in the top-level directory as build artifacts, making them
available in the Artifacts tab of each individual test step.

Signed-off-by: Christopher Maier <cmaier@chef.io>