[BUG] Update actix-web calling conventions
Per [the migration guide][1] this wraps `AppState` as `Data<AppState>`
to prevent panics in the HTTP gateway.

The rest of the code currently assumes the data is returned in this
way; future refactorings may allow us to drop the `Data` wrapper
entirely, but this is the minimal change for now.

Actix 3.0 was introduced in #7949 / #7923.

[1]: https://github.com/actix/actix-web/blob/master/MIGRATION.md#300

Signed-off-by: Christopher Maier <cmaier@chef.io>