1.6.175 Documentation, plus docs generation script update
The new `hab sup secret generate-tls` subcommand broke our document
generation script, because the script assumed that subcommands would
only contain word characters, and not hyphens.

Now, it just looks for a string of non-space characters, and goes on
with life.

The documentation string for that command was also shortened, to
ensure the output remained on one line.

Signed-off-by: Christopher Maier <cmaier@chef.io>