[FIX] Minimal change to enable non-root containers again
With the changes in #7887 for how key cache directories were handled,
we inadvertently broke non-root containers. The Supervisor creates a
cache directory explicitly when it starts (previously, it was
implicitly created). However, when running as a non-root user, there
isn't a suitable directory present.

Here, we modify the Dockerfile template to create a cache directory
for the non-root Supervisor to use.

Further cleanup and reconsideration of this pattern should be done at
some point.

Signed-off-by: Christopher Maier <cmaier@chef.io>