busybox and cacerts aren't really "base" packages
When exporting a container or tarball, we allow you to specify a
`--base-pkgs-channel`, for pulling the Launcher, Supervisor, and CLI
packages from. This also pulls busybox and cacerts from the same
channel.

Given how we run our pipelines and validation tests, we want to
specify a non-stable channel from which to pull these
packages. However, we never care about busybox or cacerts, and
shouldn't have to put them into an release channel just to validate
things.

This commit now always pulls busybox and cacerts from the stable
channel when exporting.

Signed-off-by: Christopher Maier <cmaier@chef.io>