Specify winapi version more tightly
Occasionally whenever we try to update individual dependencies, we'll
come across spurious conflicts around the `winapi` crate, where it
always tries to resolve to version 0.2.8, which at this point is
several years old.

It appears that this is arising from our broad usage of "*" dependency
constraints, along with our use of an old fork of the `ipc-channel`
crate to get support for Windows.

Here, we pin our other usage of `winapi` to modern versions, which in
turn helps keep other dependency updates sane.

(Note that this shouldn't really affect what version of `winapi` we've
actually been using in these crates; it simply codifies our
expectations more clearly.)

Signed-off-by: Christopher Maier <cmaier@chef.io>