Lock lazy_static dependency
Similar to what happened in #7976, but with the `lazy_static`
crate. Sometimes when we update other dependencies, we can fall back
to older `lazy_static` versions because we rely on "*" versioning a
bit too liberally.

(Again, this doesn't materially change anything in the code; it merely
expresses our constraints clearly.)

Signed-off-by: Christopher Maier <cmaier@chef.io>