[CI] Move documentation generation to release pipeline
To harmonize with the new company-wide way of managing our
documentation updates, we are changing how we generate our
documentation.

Previously, once a release had been promoted to stable, we generated
two pieces of dynamically-updatable documentation, created a PR for
them, and then merged them back into the repository.

Now, we generate these docs during our release pipeline, for every
potential release candidate. Rather than storing the documentation
back in the repository, we save them in a tarball (specifically,
`generated-documentation-${VERSION}.tar.gz`) in our S3 bucket
with the rest of the artifacts we generate. Later on, this artifact
will be used to assemble the final documentation that will actually be
published (that part of the process is owned by release engineering;
this PR simply makes what changes are needed to enable their work).

Since we are no longer committing generated content to the repository,
and in order to prevent future confusion, the content of the existing
documentation files has been replaced with a placeholder comment
explaining the situation.

Signed-off-by: Christopher Maier <cmaier@chef.io>