No longer assert that LOCK file PID holds the lock
Originally on Linux, this PR attempted to verify that the process of
the PID in the Supervisor's LOCK file actually held a file lock on the
LOCK file. To do this, we inspected `/proc/locks` searching for an
appropriate lock held by the process on a file with the same inode as
the LOCK file (technically, it should be inode + the device). However,
testing uncovered some inconsistencies in this approach in our CI
environment. This manifested as a different inode being present in
`/proc/locks` than returned from a `stat` call on the file itself.

Though the exact cause has not yet been determined, it is hypothesized
that it may have something to do with differences in inode handling in
overlay filesystems. The error only appears when running tests under
Docker, and only in our CI worker nodes. Given the fact that this
behavior was introduced only in this PR, as well as the possibility of
misbehavior in real-world deployments, we're removing this feature
until such time as we can implement it in a more robust way.

Signed-off-by: Christopher Maier <cmaier@chef.io>