Templated init hook now causes restarts
This commit changes the interaction of templating with restarts.
Init hooks now cause a restart if an application config change causes
the hook content to differ.

The intuitive reasoning for this is that any change that could affect
the service's runtime behaviour should cause a restart so that all
running services are consistent with the current hooks and config.

Signed-off-by: Johny Jose <johny.jose@progress.com>