introduces HAB_STUDIO_INSTALL_PKGS in studio
This env variable specifics a ':' separated list of package artifacts
that have to be installed on entering the studio. This forces habitat
to use the newer locally installed package instead of the latest remote
package version.

This ability is critical when doing package refreshes where we build
newer versions of all deps and do a studio build for each package
prioir to pushing them all to a remote builder.

Signed-off-by: Johny Jose <johny.jose@progress.com>