fixes rolling update leader resetting pkg_incarnation
During rolling updates, the update worker relies on the value of the
service group's "pkg_incarnation" to determine if an update has
occurred. If a leader in the group has a higher pkg_incarnation than
the followers, the followers will update to the leader's package.

This commit addresses a bug that caused a leader to reset it's
pkg_incarnation to 0 in certain scenarios, resulting in followers
remaining in a perpetual state of waiting for the leader to catch up
to the latest pkg_incarnation.

Signed-off-by: Johny Jose <johny.jose@progress.com>