Initial support: clap-v4 for `pkg-export-container`
Using clap-v4 APIs for the `hab-pkg-export-container` binary to mainly
demonstrate the usage of clap-v4.

Basic tests pass and basic CLI usage works eg.

```
hab-pkg-export-container core/redis
```

`--engine` was not getting correctly parsed as `EngineKind` but was
getting parsed as a `String`. This was happening because `value_parser`
set initially was getting overwritten (in the original code).

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>