`pkg-export-container`: Remove `pub` visibilty
A number of `struct`s , `enum`s and functions inside different modules
were having a `pub` visibility. A lot of these are implementation
details that need not be exposed to the caller (thus allowing us to
refactor the code without affecting the user expected behavior),

Made a single `cli_driver` public API for the `pkg_export_container`
lib crate which will be called by the binary.

There is possible opportunity for refactoring further, but we will keep
the changes small.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>