Removed dependency on crate `env`
`components/builder-api-client` had a dependency on crate `env` and
this crate is possibly yanked from `crates.io`, this resulted in builds
starting to fail.

Removed this dependency as this is an old dependency lying around and
not used anywhere in the package.

Did a very basic `cargo build` and `cargo test`

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>