Fixed issue with e2e test failing for `pkg-export-container`
In the handling of `cli` for `-i` flag, we had used a default template
of `{{pkg_origin/pkg_name}}`. This was an incorrect template, which
caused the rendered `image_name` to be empty string. That further caused
the docker build command to fail. as the `--tag` option was
`--tag :latest` instead of `--tag <image-name>:latest`

Fixed handling the proper template, added supported templates and test
cases for that.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>