Fixes to issues found during testing
Rebase with origin/main fix conflicts

Support for refresh channel and `FEATURE_LOCAL` was missing during
rebase added that support to both CLI v2 and CLI v4.

`HabPkgIdentValueParser` validator was not used for `PackageIdent` which
is a struct field in many of the `pkg` commands. Started using it.

Fixed String to Pathbuf and String to PackageIdent conversions in the
validators

Removed the unused CF and Mesos exporters.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>