Store membership in memberlist only from Alive/Suspect members
If we receive member list in `Ping` or `Ack` from a `Confirmed` member,
likely that member was in a network partition and came back, this causes
the member in network partition to overwhelm the ring with `Confirmed`
supervisors that takes a while to converge.

This would also mean for a node emerging out of network partition, it
has to receive at-least two messages (Ping and/or Ack) from another
member before that member's membership list can be accepted and stored.
This might be a small extra time to converge *after* network partition,
but is better than the entire ring diverging completely.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>