Added a new SWIM protocol message type: `ProbePing`
A `ProbePing` message will be sent to another supervisor if we receive a
`Confirmed` message from another supervisor. This message will be used
to make sure we do not blindly accept a supervisor to be `Confirmed`
based on what some other node has to say. This is particularly important
if a (set of) node(s) enter into partition and come out of partition,
then they would think other supervisors are `Confirmed` and spread that
'rumor`, resulting in otherwise `Alive` supervisors to be construed as
`Confirmed` by some supervisors.

This message will check a supposedly `Confirmed` supervisor with a
`ProbePing` message and if there is a response to that `ProbePing` the
supervisor is not considered `Confirmed`.

The `ProbePing` message is sent in the `outbound` loop. The way this is
done is we maintain a set of members to be `ProbePing`ed in a
`probe_list` inside the `Server` struct.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>