`mark_sender_alive` API
The `insert_member_mlw_rhw` API is supposed to insert the member in the
member list (note this is different from `insert_member_from_rumor`).
But due to the way we use `Incarnation`, the `Alive` health insertion is
not successful. We are using this new API for marking sender as `Alive`.

This can be better handled by `insert_member` but there are way too many
test cases around that logic, which is not very wise to change at the
moment.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>