Update Election algorithm & fix failing test case
If a node reboots while it's election is in `Running` state and during
the interval, the election is `Finished`. The rebooted node was not able
to join the finished election. If the election is `Finished` for the
same term.

If we receive a rumor that says election is not `Finished` while our
state of the election is `Finished` and we steal their vote and make a
`hot` rumor so that the other member gets a chance to `finish` their
election.

We need to *restart* election on all the partitioned followers (so that
each one of them increases term by 1) and then they can lead to election
`Finished` state.

This is what will happen in real world where the followers will
determine the leader is `Confirmed` and independently update their
*term* and start a new election.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>