Another update to Election algorithm
If a node reboots while it's election is in `Running` state and during
the interval, the election is `Finished`. The rebooted node was not able
to join the finished election. If the election is `Finished` for the
same term.

If we receive a rumor that says election is not `Finished` while our
state of the election is `Finished` and we steal their vote and make a
`hot` rumor so that the other member gets a chance to `finish` their
election.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>