Fixed issues with indices in `eachAlive` helper
The `first`, `last` etc. indices were not properly assigned in the
latest porting to new handlebars API. Fixed the issue (In the old code
we had to use the explicit `@first` , `@last` block local variables. In
the current APIs the block local variables should simply be named
`first`, `last` instead.

Also added a unit test to handle for this scenario

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>