Fix `pkg exec` cmd to treat args without `--`
The release pipeline was failing because the `pkg exec` command would
not accept the arguments to the command without a `--` separating the
arguments from the command. This caused the `pkg build` to fail for
several components.

Modified the `PkgExecOptions` to utilize a flattened `CommandArgs`
struct that is treated as the last argument.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>