Harmonizing `Origin` usage across different CLI
Different validators were used in different CLIs for the same Object
type `Origin`. Also, at a number of places we were using `String` in
place of the semantic `Origin` object in a few sub commands. Also,
whenever the 'origin' parameter was optional, we were not using the one
provided by `cli.toml` if set. Harmonized all that usage.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>