Updates end-to-end/test_external_binaries.ps1
Clap has an explicit error code of 2 to indicate that there was an issue
with parsing the flags, options, and arguments given to a program. This
test was was explicitly checking for exit code of 1. Since 1 is the
generic exit code, 2 as an exit code was intentionally added to clap,
and we have no explicit gurantees or doc around exit codes for this
command  we're updating the test to '-Not -Be 0' instead of '-Be 1'.

Signed-off-by: Jason Heath <jason.heath@progress.com>