Updates test/end-to-end/test_license.ps1
This test was calling `hab svc load --version` in order to test the
--version with a subcommand / binary other than hab. In updating to clap
v4 our practice seems to be set 'disable_version_flag=true' past the
second level. That is to day that `hab svc --version` works but the
version currently used does not.  To resolve this we have just dropped
the load subcommand from the call in the test.

Signed-off-by: Jason Heath <jason.heath@progress.com>