Requires --scaffolding to add non-placheolder scaffolding
The templates have a switch in them that wasn't being respected. In the
template_plan.sh linux version we find this version of it

|  {{ #if scaffolding_ident }}
|  pkg_scaffolding="{{ scaffolding_ident }}"
|  {{ else }}
|  # pkg_scaffolding="some/scaffolding"
|  {{ /if }}

In reality 'pkg_scaffolding='core/scaffolding=ruby"' was always being
added to the plan file when 'pkg_scaffolding="some/scaffolding"' should
have been what was added unless the --scaffolding option was used.

Signed-off-by: Jason Heath <jason.heath@progress.com>