Fixes first level subcommand doc
Some of the "first level subcommand" documentation did not match the
pattern seen in other commands resulting in a difference when generating
the cli reference documentation.

So 'hab config' was generating

| Hab-config 2.0.287/20251013110823
| The Habitat Maintainers <humans@habitat.sh>
|
| Commands relating to a Service's runtime config
|
| Usage:
| Hab config <COMMAND>

instead of

| hab-config 2.0.285
| Commands relating to a Service's runtime config
|
| Usage:
| hab config <COMMAND>

Same situaton  for the sup and svc commands. All other subcommands
except for the ones relating to 'hab studio' are correct.

Signed-off-by: Jason Heath <jason.heath@progress.com>