Changes rust-zmq dep from github repo to crates.io
Hab has used a fork (https://github.com/habitat-sh/rust-zmq) of
https://github.com/erickt/rust-zmq for the reasons given in
https://github.com/habitat-sh/habitat/pull/6444. TL:DR; at the time it
seems there wasn’t a crates.io release of rust-zmq so we created a fork
to the stability of our dependency. A crates.io version exists now
(https://crates.io/crates/zmq) so this PR removes the use of our fork
and switches to using the crate. This is OK because our changes to the
fork were

- the removal of some symlinks to licenses that were break the windows
  build. This was also done in the zmq repo at a later date.
- changing the version we leveraged from 8.? to 9.2
- the addition of copilot instruction files.

While ZMQ and the downstream rust-zmq are very stable and don’t change
much this gets us back in sync with the community so if something were
to come up, most likely a security concern at this point in zmq’s life,
this will increase our ability to see it and respond.

Signed-off-by: Jason Heath <jason.heath@progress.com>