Audits and removes expr_2021 migration
https://doc.rust-lang.org/edition-guide/rust-2024/macro-fragment-specifiers.html

That url describes the thing that I personally found to be the trickiest
thing about migrating to the 2024 edition of rust. The upshot is that is
in the 2024 edition both of these could be valid declarative macro calls

  some_macro!( _ );
  some_macro!( const { 1 + 1 } );

The declarative macro would of course have to have supporting match arms
but if the macro did have such arms that's legal in the 2024 edition and
illegal in earlier editions of rust.

I didn't find anything that matched the signature associated with const
block expressions in our rust code (but there are stmts that match the
regex in some Javascript).

The underscore expression was much more annoying because that meant
finding all of the macros that the cargo fix --edition changed to
expr_2021 and auditing them for use of _ when they were called.  I found
no such instances.

Thus, in this commit, all instsance of expr_2021 were converted back to
expr.

Signed-off-by: Jason Heath <jason.heath@progress.com>