Deals with a pair of RUSTSEC issues
First, allow lists https://rustsec.org/advisories/RUSTSEC-2025-0134
which is ultimatly a change from https://crates.io/crates/rustls-pemfile
which has been listed as unmainttained to
https://crates.io/crates/rustls-pki-types as per
https://github.com/rustls/pemfile/issues/61.

The reason for the allow listing is that we can't completely purge the
crate from habitat until several crates in the AWS stack are updated.
As a representative example see
https://github.com/awslabs/aws-sdk-rust/issues/1390.  I suspect that
this will likely be fixed soon.

Second, Flate2 yanked 1.1.6 and 1.1.7 so this downgrades us back to
1.1.5.  (https://crates.io/crates/flate2/versions)

Signed-off-by: Jason Heath <jason.heath@progress.com>