Updates run_clippy.sh for "our current rust context"
There is A LOT to this....

When looking to upgrade to 1.91.1 there were issues with running cargo
clippy due to issues with out manipulation of the dynamic interpreter.
The issue was that when clippy executed and compiled build scripts the
executables were compiled with the system interpreter instead of the
interpreter that we wanted from the base-2025 channel. The issue
presented itself as glibc mismatches as the base-2025 channel uses glibc
2.41 while my system glibc happens to be 2.36. Welcome to glibc hell
dear reader....

So, back down the rabbit hole with me and much research into ELF, it's
dynamic interpreter and runpath, and settings for cargo and rustc and
gcc, oh my....

Here's what all of that turns into.

We don't need to patchelf in this script anymore because we've pushed
the parts we really need to down into our habitatized rust pkgs.  Nixed
all that.

Because we're many years down the road from when this script was created
we also don't need to install rust in order to get clippy. Nixed all
that.

We also don't really want to play LD_LIBRARY_PATH games here for
gcc-base because they're really best played elsewhere.  The first place
begin down in our habitat packaging of rust which takes care of most
things.

However, this doesn't resolve the core issue with the environment in
which build scripts are compiled.  The resulant executables need to have
the our preferred glibc and gcc-base libraries available to them and
after much experimentation it seems best to manipulate the RUSTFLAGS
environment variable here in this script to set the ELF interpreter to
the one we're relying on and to add our preferred gcc-base to the
runpath of the build script's executable.  And this "keeps the
environment cleaner" as this targets the build scripts and doesn't mess
with LD_LIBRARY_PATH beyond what we need.

So, yeah.  That was a lot.  Thanks for reading.

Signed-off-by: Jason Heath <jason.heath@progress.com>