Updates .expeditor/scripts/verify/run_clippy.sh
When upgrading to 1.91.1 there were issues running cargo clippy due to
issues with our manipulation of the linker/dynamic interpreter.  The
issue was that when clippy executed and compiled build scripts the
executables were being compiled with the system interpreter instead of
the interpreter we wanted from the base-2025 channel. The issue
presented itself as symbols missing from glibc as base-2025 uses glibc
2.41 while my system glibc happens to be 2.36.

Here's what that turns into.

We don't need to patchelf in this script anymore because we've pushed
some of that down into our habitatized rust pkgs.  There's also a better
way to address it for other places.  Nixed all that.

Because we're many years down the road from when this script was created
we also don't need to rustup in order to get clippy. Nixed all that.

We can also play LD_LIBRARY_PATH games slightly better, maybe?  Starting
out LD_LIBRARY_PATH is managed to some extent down in our down rust
habitat packaging now.  However, even that doesn't resolve the core
issue with the environment in which build scripts are compiled.  To that
extent our preferred gcc-base library is added here.  In addition the
RUSTFLAGS are configured in order to set the dynamic interpreter our
preferred glibc more pervasively than the previous patchelf technique.

Signed-off-by: Jason Heath <jason.heath@progress.com>