Removes usage of bincode by launcher-client
Dependabot raised https://github.com/habitat-sh/habitat/pull/10102 and
it ultimately revealed a world of drama with the bincode project.

First, the lastest release seems to have itentionally broken
compilation.

```text
error: https://xkcd.com/2347/
 --> /root/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/bincode-3.0.0/src/lib.rs:1:1
  |
1 | compile_error!("https://xkcd.com/2347/");
  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

error: could not compile `bincode` (lib) due to 1 previous error
```

Which led to research which turns up both github (archived and
effectively blanked out in August) and sourcehut repos (apparently
terminated now in Dec) both of which seem to have been intentionally
wrecked due to anger over having been doxxed.

https://github.com/bincode-org/bincode
https://git.sr.ht/~stygianentity/bincode
https://crates.io/crates/bincode

It's fairly easy to find some discussion online if you would like.

The wincode dependency I've at least initially replaced it with was one
of the alternatives recommend by the bincode team in their last posting
to crates.io.  It was created and is maintained by crypto company in
service of their overall stack. Swapping it out things compiled.
Raising this PR now as the easiest way to see how it does with the
verify pipeline.

https://docs.rs/wincode/latest/wincode/

If we want to do a deeper review of alternatives we can do that.

Signed-off-by: Jason Heath <jason.heath@progress.com>