`HAB_STUDIO_INSTALL_PKGS` for default studio
Similar to the `bootstrap` support, this support is added for the default studio.
For the normal (default) invocation of the studio, this should typically be
invoked in the dev mode to install local dependencies, the usual execution mode
would install packages from the builder and this env variable is not required.

Typical use-case is building packages using `hab-auto-build`.

Signed-off-by: Abhijit Gadgil <agadgil@progress.com>